refactor: change the way in which we vendor psiphon
This is an experiment. I have no way to be sure it's WAI until I tag
a new version and try to use it from ooni/probe-cli (well, maybe there
is another way, but I don't know this way).

Here the idea is that we'll try to use most dependencies directly
and we'll just vendor the problematic Psiphon dependencies.

Reference issue: https://github.com/ooni/probe/issues/1894